Updated code to comply with python 3.7
Code will now work with all versions of python.

This script now fully works and outputs a list of users with their total likes.

There is a known issue with larger profiles that I believe could be mitigated through the use of the RightPagination button once the first picture is opened